Moved calibration module from core to utils.
