+++ stellargraph/__init__.py
@@ -11,8 +11,8 @@ from .version import __version__
 # Top-level imports
 from stellargraph.core.graph import StellarGraph, StellarDiGraph
 from stellargraph.core.schema import GraphSchema
-from stellargraph.core.calibration import TemperatureCalibration, IsotonicCalibration
-from stellargraph.core.calibration import (
+from stellargraph.utils.calibration import TemperatureCalibration, IsotonicCalibration
+from stellargraph.utils.calibration import (
     plot_reliability_diagram,
     expected_calibration_error,
 )
